+++ index.html
@@ -229,7 +229,7 @@
 				<div class="row text-center">
 					<div class="col form-group col-12 col-md-8 offset-md-2">
 						<div class="input-group">
-							<button id="genLink" type="button" class="btn btn-outline-primary">Generate Share Link</button>
+							<button id="genLink" type="button" class="btn btn-outline-primary" onclick="generateShareLink()">Generate Share Link</button>
 							<input id="shareLink" class="form-control" placeholder="No link" type="text" readonly>
 							<span id="linkValid" class="input-group-text"></span>
 						</div>
